optimizations

+++ magasine/models.py
@@ -34,7 +34,7 @@ class Product(models.Model):
 class OrderProduct(models.Model):
     author = models.ForeignKey(CustomUser, on_delete=models.CASCADE)
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
-    quantity = models.FloatField()
+    quantity = models.IntegerField()
 
     def __str__(self):
         return self.product.modelname
